requirements.txt
playwright
python-dotenv
flask
flask-socketio
redis
requests
gunicorn
eventlet 